+++ sdk/webpubsub/azure-messaging-webpubsubservice/azure/messaging/webpubsubservice/_patch.py
@@ -24,92 +24,23 @@
 # IN THE SOFTWARE.
 #
 # --------------------------------------------------------------------------
-import typing
+from typing import Any, TYPE_CHECKING, Optional, Union, Awaitable
 import jwt
 import six
-from datetime import datetime, timedelta, tzinfo
-from typing import TYPE_CHECKING
-import importlib
+from datetime import datetime, timedelta
 
-from ._version import VERSION
-from ._web_pub_sub_service_client import WebPubSubServiceClient as GeneratedWebPubSubServiceClient
+from ._web_pub_sub_service_client import WebPubSubServiceClient as WebPubSubServiceClientGenerated
+from ._operations._patch import _UTC_TZ
 
-from msrest import Deserializer, Serializer
-from azure.core.pipeline import policies
-from azure.core import PipelineClient
-from azure.core.configuration import Configuration
+from azure.core.pipeline import PipelineRequest
 from azure.core.pipeline.policies import SansIOHTTPPolicy, ProxyPolicy
 from azure.core.credentials import AzureKeyCredential
-from azure.core.pipeline import PipelineResponse
-from azure.core.pipeline.transport import HttpResponse
-from azure.core.rest import HttpRequest
-from azure.core.tracing.decorator import distributed_trace
-from msrest import Serializer
-
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
-    from typing import Any, Callable, Dict, Optional, TypeVar, Union
-    from ._operations._operations import JSONType
-
-    T = TypeVar("T")
-    ClsType = Optional[Callable[[PipelineResponse[HttpRequest, HttpResponse], T, Dict[str, Any]], Any]]
-
-if TYPE_CHECKING:
-    # pylint: disable=unused-import,ungrouped-imports
-    from typing import Type, TypeVar, Any, Union, Dict
-
     from azure.core.credentials import TokenCredential
-    from azure.core.pipeline import PipelineRequest
 
-
-class _UTC_TZ(tzinfo):
-    """from https://docs.python.org/2/library/datetime.html#tzinfo-objects"""
-
-    ZERO = timedelta(0)
-
-    def utcoffset(self, dt):
-        return self.__class__.ZERO
-
-    def tzname(self, dt):
-        return "UTC"
-
-    def dst(self, dt):
-        return self.__class__.ZERO
-
-
-def _get_token_by_key(endpoint: str, hub: str, key: str, **kwargs: Any) -> str:
-    """build token with access key.
-
-    :param endpoint:  HTTPS endpoint for the WebPubSub service instance.
-    :type endpoint: str
-    :param hub: The hub to give access to.
-    :type hub: str
-    :param key: The access key
-    :type hub: str
-    :keyword dict[str, any] jwt_headers: Any headers you want to pass to jwt encoding.
-    :returns: token
-    :rtype: str
-    """
-    audience = "{}/client/hubs/{}".format(endpoint, hub)
-    user = kwargs.pop("user_id", None)
-    ttl = timedelta(minutes=kwargs.pop("minutes_to_expire", 60))
-    roles = kwargs.pop("roles", [])
-
-    payload = {
-        "aud": audience,
-        "iat": datetime.now(tz=_UTC_TZ()),
-        "exp": datetime.now(tz=_UTC_TZ()) + ttl,
-    }
-    if user:
-        payload["sub"] = user
-    if roles:
-        payload["role"] = roles
-
-    return six.ensure_str(jwt.encode(payload, key, algorithm="HS256", headers=kwargs.pop("jwt_headers", {})))
-
-
-def _parse_connection_string(connection_string: str, **kwargs: Any) -> JSONType:
+def _parse_connection_string(connection_string: str, **kwargs: Any) -> Any:
     for segment in connection_string.split(";"):
         if "=" in segment:
             key, value = segment.split("=", 1)
@@ -140,9 +71,9 @@ class JwtCredentialPolicy(SansIOHTTPPolicy):
         self,
         credential: AzureKeyCredential,
         *,
-        user: typing.Optional[str] = None,
-        origin_endpoint: typing.Optional[str] = None,
-        reverse_proxy_endpoint: typing.Optional[str] = None,
+        user: Optional[str] = None,
+        origin_endpoint: Optional[str] = None,
+        reverse_proxy_endpoint: Optional[str] = None,
     ) -> None:
         """Create a new instance of the policy associated with the given credential.
 
@@ -156,7 +87,7 @@ class JwtCredentialPolicy(SansIOHTTPPolicy):
         self._original_url = origin_endpoint
         self._reverse_proxy_endpoint = reverse_proxy_endpoint
 
-    def on_request(self, request: PipelineRequest) -> typing.Union[None, typing.Awaitable[None]]:
+    def on_request(self, request: PipelineRequest) -> Union[None, Awaitable[None]]:
         """Is executed before sending the request from next policy.
 
         :param request: Request to be modified before sent from next policy.
@@ -207,71 +138,24 @@ class ApiManagementProxy(ProxyPolicy):
         if self._endpoint and self._reverse_proxy_endpoint:
             request.http_request.url = request.http_request.url.replace(self._endpoint, self._reverse_proxy_endpoint)
 
+class WebPubSubServiceClientBase:
+    """Base class for init"""
 
-class WebPubSubServiceClientConfiguration(Configuration):
-    """Configuration for WebPubSubServiceClient.
+    def __init__(self, endpoint: str, hub: str, credential, **kwargs) -> None:
+        if kwargs.get("port") and endpoint:
+            endpoint = endpoint.rstrip("/") + ":{}".format(kwargs.pop("port"))
+        kwargs["origin_endpoint"] = endpoint
+        if isinstance(credential, AzureKeyCredential):
+            kwargs["authentication_policy"] = JwtCredentialPolicy(
+                credential,
+                user=kwargs.get("user"),
+                origin_endpoint=kwargs.get("origin_endpoint"),
+                reverse_proxy_endpoint=kwargs.get("reverse_proxy_endpoint")
+            )
+        kwargs["proxy_policy"] = kwargs.pop("proxy_policy", ApiManagementProxy(**kwargs))
+        super().__init__(endpoint=endpoint, hub=hub, credential=credential, **kwargs)
 
-    Note that all parameters used to create this instance are saved as instance
-    attributes.
-
-    :param hub: Target hub name, which should start with alphabetic characters and only contain alpha-numeric characters or underscore.
-    :type hub: str
-    :param endpoint: HTTP or HTTPS endpoint for the Web PubSub service instance.
-    :type endpoint: str
-    :param credential: Credential needed for the client to connect to Azure.
-    :type credential: Union[~azure.core.credentials.TokenCredential, ~azure.core.credentials.AzureKeyCredential]
-    :keyword api_version: Api Version. The default value is "2021-10-01". Note that overriding this default value may result in unsupported behavior.
-    """
-
-    def __init__(
-        self, hub: str, endpoint: str, credential: Union[TokenCredential, AzureKeyCredential], **kwargs: Any
-    ) -> None:
-        super().__init__(**kwargs)
-        api_version = kwargs.pop("api_version", "2021-10-01")  # type: str
-
-        if hub is None:
-            raise ValueError("Parameter 'hub' must not be None.")
-        if endpoint is None:
-            raise ValueError("Parameter 'endpoint' must not be None.")
-        if credential is None:
-            raise ValueError("Parameter 'credential' must not be None.")
-
-        self.hub = hub
-        self.endpoint = endpoint
-        self.credential = credential
-        self.api_version = api_version
-        self.credential_scopes = kwargs.pop("credential_scopes", ["https://webpubsub.azure.com/.default"])
-        kwargs.setdefault("sdk_moniker", "messaging-webpubsubservice/{}".format(VERSION))
-        self._configure(**kwargs)
-
-    def _configure(
-        self, **kwargs  # type: Any
-    ):
-        # type: (...) -> None
-        self.user_agent_policy = kwargs.get("user_agent_policy") or policies.UserAgentPolicy(**kwargs)
-        self.headers_policy = kwargs.get("headers_policy") or policies.HeadersPolicy(**kwargs)
-        self.proxy_policy = kwargs.get("proxy_policy") or ApiManagementProxy(**kwargs)
-        self.logging_policy = kwargs.get("logging_policy") or policies.NetworkTraceLoggingPolicy(**kwargs)
-        self.http_logging_policy = kwargs.get("http_logging_policy") or policies.HttpLoggingPolicy(**kwargs)
-        self.retry_policy = kwargs.get("retry_policy") or policies.RetryPolicy(**kwargs)
-        self.custom_hook_policy = kwargs.get("custom_hook_policy") or policies.CustomHookPolicy(**kwargs)
-        self.redirect_policy = kwargs.get("redirect_policy") or policies.RedirectPolicy(**kwargs)
-        self.authentication_policy = kwargs.get("authentication_policy")
-        if self.credential and not self.authentication_policy:
-            if isinstance(self.credential, AzureKeyCredential):
-                self.authentication_policy = JwtCredentialPolicy(
-                    self.credential,
-                    user=kwargs.get("user"),
-                    origin_endpoint=kwargs.get("origin_endpoint"),
-                    reverse_proxy_endpoint=kwargs.get("reverse_proxy_endpoint"),
-                )
-            else:
-                self.authentication_policy = policies.BearerTokenCredentialPolicy(
-                    self.credential, *self.credential_scopes, **kwargs
-                )
-
-
-class WebPubSubServiceClient(GeneratedWebPubSubServiceClient):
+class WebPubSubServiceClient(WebPubSubServiceClientBase, WebPubSubServiceClientGenerated):
     """WebPubSubServiceClient.
 
     :param endpoint: HTTP or HTTPS endpoint for the Web PubSub service instance.
@@ -287,18 +171,11 @@ class WebPubSubServiceClient(GeneratedWebPubSubServiceClient):
     """
 
     def __init__(
-        self, endpoint: str, hub: str, credential: Union[TokenCredential, AzureKeyCredential], **kwargs: Any
+        self, endpoint: str, hub: str, credential: Union["TokenCredential", AzureKeyCredential], **kwargs: Any
     ) -> None:
-        if kwargs.get("port") and endpoint:
-            endpoint = endpoint.rstrip("/") + ":{}".format(kwargs.pop("port"))
-        kwargs["origin_endpoint"] = endpoint
-        _endpoint = "{Endpoint}"
-        self._config = WebPubSubServiceClientConfiguration(hub=hub, endpoint=endpoint, credential=credential, **kwargs)
-        self._client = PipelineClient(base_url=_endpoint, config=self._config, **kwargs)
-
-        self._serialize = Serializer()
-        self._deserialize = Deserializer()
-        self._serialize.client_side_validation = False
+        super().__init__(
+            endpoint=endpoint, hub=hub, credential=credential, **kwargs
+        )
 
     @classmethod
     def from_connection_string(cls, connection_string: str, hub: str, **kwargs: Any) -> "WebPubSubServiceClient":
@@ -316,56 +193,7 @@ class WebPubSubServiceClient(GeneratedWebPubSubServiceClient):
         credential = AzureKeyCredential(kwargs.pop("accesskey"))
         return cls(hub=hub, credential=credential, **kwargs)
 
-    @distributed_trace
-    def get_client_access_token(self, **kwargs: Any) -> JSONType:
-        """Build an authentication token.
-
-        :keyword user_id: User Id.
-        :paramtype user_id: str
-        :keyword roles: Roles that the connection with the generated token will have.
-        :paramtype roles: list[str]
-        :keyword minutes_to_expire: The expire time of the generated token.
-        :paramtype minutes_to_expire: int
-        :keyword dict[str, any] jwt_headers: Any headers you want to pass to jwt encoding.
-        :returns: JSON response containing the web socket endpoint, the token and a url with the generated access token.
-        :rtype: JSONType
-
-        Example:
-        >>> get_client_access_token()
-        {
-            'baseUrl': 'wss://contoso.com/api/webpubsub/client/hubs/theHub',
-            'token': 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ...',
-            'url': 'wss://contoso.com/api/webpubsub/client/hubs/theHub?access_token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ...'
-        }
-        """
-        endpoint = self._config.endpoint.lower()
-        if not endpoint.startswith("http://") and not endpoint.startswith("https://"):
-            raise ValueError(
-                "Invalid endpoint: '{}' has unknown scheme - expected 'http://' or 'https://'".format(endpoint)
-            )
-
-        # Ensure endpoint has no trailing slash
-        endpoint = endpoint.rstrip("/")
-
-        # Switch from http(s) to ws(s) scheme
-        client_endpoint = "ws" + endpoint[4:]
-        hub = self._config.hub
-        client_url = "{}/client/hubs/{}".format(client_endpoint, hub)
-        jwt_headers = kwargs.pop("jwt_headers", {})
-        if isinstance(self._config.credential, AzureKeyCredential):
-            token = _get_token_by_key(endpoint, hub, self._config.credential.key, jwt_headers=jwt_headers, **kwargs)
-        else:
-            token = super(WebPubSubServiceClient, self).get_client_access_token(**kwargs).get("token")
-
-        return {
-            "baseUrl": client_url,
-            "token": token,
-            "url": "{}?access_token={}".format(client_url, token),
-        }
-
-    get_client_access_token.metadata = {"url": "/api/hubs/{hub}/:generateToken"}  # type: ignore
-
+__all__ = ["WebPubSubServiceClient"]
 
 def patch_sdk():
-    curr_package = importlib.import_module("azure.messaging.webpubsubservice")
-    curr_package.WebPubSubServiceClient = WebPubSubServiceClient
+    pass
